+++ src/aioway/instrs/layers/actives.py
@@ -2,7 +2,7 @@
 
 from torch import nn
 
-from ..nn import NnInstr, nn_instr_dcls
+from ..nn import NnInstr, instr_dcls
 
 __all__ = [
     "ReLU",
@@ -17,63 +17,63 @@ __all__ = [
 ]
 
 
-@nn_instr_dcls
+@instr_dcls
 class ReLU(NnInstr):
     "Applies the rectified linear unit function element-wise."
 
     NN = nn.ReLU
 
 
-@nn_instr_dcls
+@instr_dcls
 class ReLU6(NnInstr):
     "Applies the ReLU6 function element-wise."
 
     NN = nn.ReLU6
 
 
-@nn_instr_dcls
+@instr_dcls
 class CELU(NnInstr):
     "Applies the CELU function element-wise."
 
     NN = nn.CELU
 
 
-@nn_instr_dcls
+@instr_dcls
 class GELU(NnInstr):
     "Applies the GELU function element-wise."
 
     NN = nn.GELU
 
 
-@nn_instr_dcls
+@instr_dcls
 class Sigmoid(NnInstr):
     "Applies the Sigmoid function element-wise."
 
     NN = nn.Sigmoid
 
 
-@nn_instr_dcls
+@instr_dcls
 class Tanh(NnInstr):
     "Applies the Tanh function element-wise."
 
     NN = nn.Tanh
 
 
-@nn_instr_dcls
+@instr_dcls
 class Softmin(NnInstr):
     "Applies the Softmin function to an n-dimensional input Tensor."
 
     NN = nn.Softmin
 
 
-@nn_instr_dcls
+@instr_dcls
 class Softmax(NnInstr):
     "Applies the Softmax function to an n-dimensional input Tensor."
 
     NN = nn.Softmax
 
 
-@nn_instr_dcls
+@instr_dcls
 class LogSoftmax(NnInstr):
     "Applies the LogSoftmax function to an n-dimensional input Tensor."
