+++ src/aioway/instrs/layers/dropouts.py
@@ -4,12 +4,12 @@ import typing
 
 from torch import nn
 
-from ..nn import NnInstr, nn_instr_dcls
+from ..nn import NnInstr, instr_dcls
 
 __all__ = ["Dropout", "Dropout1d", "Dropout2d", "Dropout3d"]
 
 
-@nn_instr_dcls
+@instr_dcls
 class _BaseDropout(NnInstr):
     p: float = 0.5
     "Probability of an element to be zeroed. Default: 0.5."
@@ -18,7 +18,7 @@ class _BaseDropout(NnInstr):
     "If set to True, will do this operation in-place. Default: `False`."
 
 
-@nn_instr_dcls
+@instr_dcls
 class Dropout(_BaseDropout):
     """
     During training, randomly zeroes some of the elements
@@ -28,7 +28,7 @@ class Dropout(_BaseDropout):
     NN = nn.Dropout
 
 
-@nn_instr_dcls
+@instr_dcls
 class Dropout1d(_BaseDropout):
     """
     Randomly zero out entire channels (1D feature map).
@@ -37,7 +37,7 @@ class Dropout1d(_BaseDropout):
     NN = nn.Dropout1d
 
 
-@nn_instr_dcls
+@instr_dcls
 class Dropout2d(_BaseDropout):
     """
     Randomly zero out entire channels (2D feature map).
@@ -46,7 +46,7 @@ class Dropout2d(_BaseDropout):
     NN = nn.Dropout2d
 
 
-@nn_instr_dcls
+@instr_dcls
 class Dropout3d(_BaseDropout):
     """
     Randomly zero out entire channels (3D feature map).
