+++ src/aioway/instrs/layers/embs.py
@@ -4,12 +4,12 @@ import typing
 
 from torch import nn
 
-from ..nn import NnInstr, nn_instr_dcls
+from ..nn import NnInstr, instr_dcls
 
 __all__ = ["Embedding"]
 
 
-@nn_instr_dcls
+@instr_dcls
 class _BaseEmbedding(NnInstr):
     num_embeddings: int
     "The size of the dictionary of embeddings."
@@ -25,7 +25,7 @@ class _BaseEmbedding(NnInstr):
             raise ValueError(f"{self.embedding_dim=} <= 0.")
 
 
-@nn_instr_dcls
+@instr_dcls
 class Embedding(_BaseEmbedding):
     """
     A simple lookup table that stores embeddings of a fixed dictionary and size.
@@ -34,7 +34,7 @@ class Embedding(_BaseEmbedding):
     NN = nn.Embedding
 
 
-@nn_instr_dcls
+@instr_dcls
 class EmbeddingBag(_BaseEmbedding):
     """
     Compute sums or means of 'bags' of embeddings,
