+++ src/aioway/instrs/losses.py
@@ -4,7 +4,7 @@ import typing
 
 from torch import nn
 
-from ..nn import NnInstr, nn_instr_dcls
+from .nn import NnInstr, instr_dcls
 
 __all__ = [
     "BaseLoss",
@@ -22,7 +22,7 @@ __all__ = [
 _REDUCTION = frozenset(["none", "mean", "sum"])
 
 
-@nn_instr_dcls
+@instr_dcls
 class BaseLoss(NnInstr):
     """
     Creates a criterion that measures the mean absolute error (MAE)
@@ -30,7 +30,7 @@ class BaseLoss(NnInstr):
     """
 
 
-@nn_instr_dcls
+@instr_dcls
 class L1Loss(BaseLoss):
     """
     Creates a criterion that measures the mean absolute error (MAE)
@@ -40,7 +40,7 @@ class L1Loss(BaseLoss):
     NN = nn.L1Loss
 
 
-@nn_instr_dcls
+@instr_dcls
 class MSELoss(BaseLoss):
     """
     Creates a criterion that measures the mean squared error (squared L2 norm)
@@ -50,7 +50,7 @@ class MSELoss(BaseLoss):
     NN = nn.MSELoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class CrossEntropyLoss(BaseLoss):
     """
     This criterion computes the cross entropy loss between input logits and target.
@@ -59,7 +59,7 @@ class CrossEntropyLoss(BaseLoss):
     NN = nn.CrossEntropyLoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class CTCLoss(BaseLoss):
     """
     The Connectionist Temporal Classification loss.
@@ -68,7 +68,7 @@ class CTCLoss(BaseLoss):
     NN = nn.CTCLoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class NLLLoss(BaseLoss):
     """
     The negative log likelihood loss.
@@ -78,7 +78,7 @@ class NLLLoss(BaseLoss):
     NN = nn.NLLLoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class KLDivLoss(BaseLoss):
     """
     The Kullback-Leibler divergence loss.
@@ -87,7 +87,7 @@ class KLDivLoss(BaseLoss):
     NN = nn.KLDivLoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class BCELoss(BaseLoss):
     """
     Creates a criterion that measures the Binary Cross Entropy between the target and the input probabilities.
@@ -96,7 +96,7 @@ class BCELoss(BaseLoss):
     NN = nn.BCELoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class BCEWithLogitsLoss(BaseLoss):
     """
     This loss combines a Sigmoid layer and the BCELoss in one single class.
@@ -108,7 +108,7 @@ class BCEWithLogitsLoss(BaseLoss):
     NN = nn.BCEWithLogitsLoss
 
 
-@nn_instr_dcls
+@instr_dcls
 class SmoothL1Loss(BaseLoss):
     """
     Creates a criterion that uses a squared term
